daqdev/NSCLDAQ#975 NSCLDAQ SUPPORTED LOGBOOK
Testing revealed that the image offsets actually were not being
properly stored but tests missed that because we only tested
with offsets equal to zero:
*  Modified a test in notetests.cpp to display that error.
*  Figured out and fixed the problem:  offsets are size_t
   which cause the uint64_t bind to be called and that
   doesn't bind the uint64_t but, for reasons shrouded in
   memory make a null filled blob who's size is the uint64_t
   passed.  So in LogBookNote.cpp, cast the offset to an int
   which fixes that problem (limiting the notes to a 'mere'
   4 billion bytes in length.

+++ main/utilities/logbook/LogBookNote.cpp
@@ -390,7 +390,12 @@ LogBookNote::create(
         );
         insimage.bind(1, noteId);
         for (int i =0; i < images.size(); i++) {
-            insimage.bind(2, images[i].s_offset);
+            
+            // The cast to int is required below because size_t may be
+            // 64 bits and our bind for 64 bit int creates a fixed sized,
+            // null filled blob (placeholder)... is what it is..
+            
+            insimage.bind(2, (int)(images[i].s_offset));
             insimage.bind(3, images[i].s_filename.c_str(), -1, SQLITE_STATIC);
             auto img = readImage(images[i].s_filename);
             size_t      isize = img.first;
